[TASK] Split functional tests
Tests are split by a testing group. This should help understanding their
purpose and to be able to add new tests in a clear structure.
